Add script converting Excel CSV to UTF-8 encoding
When you ’Save as CSV’ from Excel, the result has windows-1251 encoding
(aka ‘ansi’), which is meaningless for most software, especially on
Mac. The quickest way to deal with this problem is to convert the data
to UTF-8. This script iterates through all the files in its working
directory, encodes them to UTF-8 and saves the correctly encoded copies
to adjacent ‘data-encoded’ folder.